epi-risk-lite/pyproject.toml
[project]
name = "epi-risk-lite"
version = "0.1.0"
description = "Clinical decision support for epistatic risk scoring"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "camelot-py[cv]>=0.11.0",
    "pdfplumber>=0.10.0",
    "xgboost>=2.0.0",
    "scikit-learn>=1.3.0",
    "joblib>=1.3.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["app/tests"]
asyncio_mode = "auto"

